+++ packages/magnitude-test/src/term-app/constants.ts
@@ -1,0 +1,22 @@
+// --- ANSI Escape Codes ---
+export const ANSI_RESET = '\x1b[0m';
+export const ANSI_BRIGHT_RED = '\x1b[91m';
+export const ANSI_BRIGHT_GREEN = '\x1b[92m';
+export const ANSI_BRIGHT_BLUE = '\x1b[94m';
+export const ANSI_GRAY = '\x1b[90m';
+export const ANSI_RED = '\x1b[31m';
+export const ANSI_BOLD = '\x1b[1m';
+export const ANSI_DIM = '\x1b[2m';
+
+// --- Box Drawing Characters ---
+export const BOX_CHARS_ROUNDED = {
+    topLeft: '╭', topRight: '╮', bottomLeft: '╰', bottomRight: '╯',
+    horizontal: '─', vertical: '│'
+};
+
+// --- Configuration ---
+export const MAX_APP_WIDTH = 100;
+export const PADDING = 2;
+
+// --- Spinner Characters ---
+export const spinnerChars = ['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '⠏'];
